Refactor: Move Page Assembly and the Cache Decorator Out
Nineteen module-level names in api_resource.py built the two
server-rendered pages: the static directory, three template paths, five
HTML fragments read at import, three asset content-hashes, fragment
injection, minification, and the two page builders. They are now
api/utils/page_rendering.py.

The `cached` decorator came out first, into api/utils/caching.py, because
three destinations use it — page assembly, get_where_clause, and the
admin import path. It exists because cachebox.cached binds its decision
at decoration time; this consults the setting per call.

STATIC_DIR is anchored on the package directory rather than counting
parents from __file__, since moving the module a level deeper is exactly
what silently repointed it at api/utils/static during this change.

TestHtmlMinification moved to test_page_rendering.py.

api_resource.py: 2423 -> 2307 lines.
